fix: Fixes bugs and adds JavaDoc to the AST RBE.

+++ src/org/jitsi/impl/neomedia/rtp/remotebitrateestimator/RemoteBitrateEstimatorWrapper.java
@@ -120,7 +120,7 @@ public class RemoteBitrateEstimatorWrapper
      * {@inheritDoc}
      */
     @Override
-    public Collection<Integer> getSsrcs()
+    public Collection<Long> getSsrcs()
     {
         return rbe.getSsrcs();
     }
@@ -129,7 +129,7 @@ public class RemoteBitrateEstimatorWrapper
      * {@inheritDoc}
      */
     @Override
-    public void removeStream(int ssrc)
+    public void removeStream(long ssrc)
     {
         rbe.removeStream(ssrc);
     }
